Intercept ACME only if we do SSL-Termination
Previously we intercepted all HTTP-Requests to /.well-known/acme-challenge/
to be able to get SSL-Certificates via let's encrypt. That prevented downstream
backends from doing let's encrypt.

With this commit we only do the intercept when it's necessary.

It's not necessary for us to do the intercept when we do not do SSL-Termination,
because in that case we do not need the certificate anyway and do not attempt
to get it.