Add wiki-test and www-test, fix issue with to many non-SSL-offloading-Hosts
The condition for the switch to the local letsencrypt-acme-challenge-folder failed because we had to many non-letsencrypt-hosts and haproxy has a limit to how many „words“ can be in one line.

With this new version we do not use one condition per non-letsencrypt-host, we use one line per letsencrypt-host. Because lines are unlimited this fixes the issue for all numbres of letsencrypt-hosts and non-letsencrypt-hosts